+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Fixed
 * <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)
-* None.
+* Upgrading files with string primary keys would result in a file where it was not possible to find the objects by primary key ([#6716](https://github.com/realm/realm-cocoa/issues/6716), since 6.0.7)
  
 ### Breaking changes
 * None.
@@ -13,7 +13,7 @@
 -----------
 
 ### Internals
-* None.
+* File format version bumped to 11.
 
 ----------------------------------------------